Move strOptions to tags/options.js
