+++ schema.js
@@ -1,4 +1,5 @@
 module.exports = require('./dist/schema').default
-module.exports.nullOptions = require('./dist/tags/options').nullOptions
+var opt = require('./dist/tags/options')
+module.exports.nullOptions = opt.nullOptions
+module.exports.strOptions = opt.strOptions
 module.exports.stringify = require('./dist/stringify').default
-module.exports.strOptions = require('./dist/stringify').strOptions
