+++ tests/doc/foldFlowLines.js
@@ -1,6 +1,6 @@
 import fold, { FOLD_FLOW, FOLD_QUOTED } from '../../src/foldFlowLines'
 import YAML from '../../src/index'
-import { strOptions } from '../../src/stringify'
+import { strOptions } from '../../src/tags/options'
 
 describe('plain', () => {
   const src = 'abc def ghi jkl mno pqr stu vwx yz\n'
